backend/requirements.txt
aiosqlite==0.22.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
APScheduler==3.11.2
beautifulsoup4==4.14.3
certifi==2026.4.22
charset-normalizer==3.4.7
click==8.3.3
colorama==0.4.6
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.136.1
greenlet==3.5.0
h11==0.16.0
idna==3.13
lxml==6.1.0
pydantic==2.13.4
pydantic_core==2.46.4
python-dotenv==1.2.2
requests==2.33.1
soupsieve==2.8.3
SQLAlchemy==2.0.49
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.2
tzlocal==5.3.1
urllib3==2.7.0
uvicorn==0.46.0
